Moved inference action to server
