+++ server/src/main/java/module-info.java
@@ -134,6 +134,7 @@ module org.elasticsearch.server {
     exports org.elasticsearch.action.fieldcaps;
     exports org.elasticsearch.action.get;
     exports org.elasticsearch.action.index;
+    exports org.elasticsearch.action.inference;
     exports org.elasticsearch.action.ingest;
     exports org.elasticsearch.action.resync;
     exports org.elasticsearch.action.search;
